interface voiture et ajoutvoiture faite

+++ src/Voiture.java
@@ -2,10 +2,10 @@ public class Voiture extends Vehicule{
 
     private float km;
     private int nbPlaces;
-    private int puissance;
+    private float puissance;
     private float prixLoc;
 
-    public Voiture(String marque, String modele, float vitesseMax, String etat, float km, int nbPlaces, int puissance, float prixLoc) {
+    public Voiture(String marque, String modele, float vitesseMax, String etat, float km, int nbPlaces, float puissance, float prixLoc) {
         super(marque, modele, vitesseMax, etat);
         this.km = km;
         this.nbPlaces = nbPlaces;
@@ -51,7 +51,7 @@ public class Voiture extends Vehicule{
         return nbPlaces;
     }
 
-    public int getPuissance() {
+    public float getPuissance() {
         return puissance;
     }
